FIX Logger
Define logger message

Resume
https://trello.com/c/0TbHApez

+++ coreapi-framework.service/src/main/java/org/sylrsykssoft/coreapi/framework/service/BaseEntityService.java
@@ -13,13 +13,15 @@ import org.sylrsykssoft.coreapi.framework.database.exception.IncorrectResultSize
 import org.sylrsykssoft.coreapi.framework.database.exception.NotFoundEntityException;
 import org.sylrsykssoft.coreapi.framework.database.repository.BaseEntityRepository;
 import org.sylrsykssoft.coreapi.framework.library.mapper.IMapperFunction;
+import org.sylrsykssoft.coreapi.framework.library.util.LoggerUtil;
+import org.sylrsykssoft.coreapi.framework.library.util.LoggerUtil.LogMessageLevel;
 
 /**
  * BaseService service.
  * 
- * @author juan.gonzalez.fernandez.jgf
- *
  * @param <T> Type class.
+ * @param <R> Resource class.
+ * @author juan.gonzalez.fernandez.jgf
  */
 public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEntityResource> implements IEntityService<T, R, Long>, IMapperFunction<T, R> {
 
@@ -30,17 +32,20 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 	public long count() {
 		return getEntityRepository().count();
 	}
-	
+
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
 	public R create(final R entity) throws NotFoundEntityException {
 		final T source = getEntityRepository().save(mapperToEntity().apply(entity));
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::create Result {}.", source);
+
 		// Convert entity to resource
 		return mapperToResource().toResource(source);
 	}
-	
+
 	/**
 	 * {@inheritDoc}
 	 */
@@ -48,19 +53,23 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 	public void delete(final R source) throws NotFoundEntityException {
 		if (source.getEntityId() != null && !existsById(source.getEntityId()))
 			throw new NotFoundEntityException();
-		
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::delete Entity {} to delete.", source);
+
 		final T entity = mapperToEntity().apply(source);
 		getEntityRepository().delete(entity);
 	}
-	
+
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
 	public void deleteAll() {
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::deleteAll.");
+
 		getEntityRepository().deleteAll();
 	}
-	
+
 	/**
 	 * {@inheritDoc}
 	 */
@@ -69,7 +78,9 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 		final Iterable<T> entities = StreamSupport.stream(sources.spliterator(), false)
 				.map(mapperToEntity()::apply)
 				.collect(Collectors.toList());
-		
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::deleteAll Delete all entities {}.", entities);
+
 		getEntityRepository().deleteAll(entities);
 	}
 
@@ -78,23 +89,32 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 	 */
 	@Override
 	public void deleteById(final Long id) throws NotFoundEntityException {
-		getEntityRepository().deleteById(id);
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::deleteById Delete entity with id {}.", id);
+
+		if (existsById(id)) {
+			getEntityRepository().deleteById(id);
+		}
 	}
-	
+
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
 	public boolean existsById(final Long id) {
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::existsById Exists entity with id {}.", id);
+
 		return getEntityRepository().existsById(id);
 	}
-	
+
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
 	public List<R> findAll() {
 		final List<T> sources = getEntityRepository().findAll();
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::findAll Found {} entries.", sources);
+
 		// Convert entity to resource
 		return sources.stream().map(mapperToResource()::toResource).collect(Collectors.toList());
 	}
@@ -106,8 +126,11 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 	public List<R> findAllByExample(final Example<R> example) {
 		final T entity = mapperToEntity().apply(example.getProbe());
 		final Example<T> exampleToFind = Example.of(entity, example.getMatcher());
-		
+
 		final List<T> sources = getEntityRepository().findAll(exampleToFind);
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::findAllByExample Found {} entries.", sources);
+
 		// Convert entity to resource
 		return sources.stream().map(mapperToResource()::toResource).collect(Collectors.toList());
 	}
@@ -119,18 +142,25 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 	public List<R> findAllByExampleSortable(final Example<R> example, final Sort sort) {
 		final T entity = mapperToEntity().apply(example.getProbe());
 		final Example<T> exampleToFind = Example.of(entity, example.getMatcher());
-		
+
 		final List<T> sources = getEntityRepository().findAll(exampleToFind, sort);
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::findAllByExampleSortable Found {} entries.",
+				sources);
+
 		// Convert entity to resource
 		return sources.stream().map(mapperToResource()::toResource).collect(Collectors.toList());
 	}
-	
+
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
 	public List<R> findAllById(final Iterable<Long> ids) {
 		final List<T> sources = getEntityRepository().findAllById(ids);
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::findAllById Found {} entries.", sources);
+
 		// Convert entity to resource
 		return sources.stream().map(mapperToResource()::toResource).collect(Collectors.toList());
 	}
@@ -141,9 +171,15 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 	@Override
 	public Optional<R> findByExample(final Example<R> example) throws NotFoundEntityException, IncorrectResultSizeException {
 		final T entity = mapperToEntity().apply(example.getProbe());
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::findByExample Finding a entry with: {}", entity);
+
 		final Example<T> exampleToFind = Example.of(entity, example.getMatcher());
-		
+
 		final Optional<T> source = getEntityRepository().findOne(exampleToFind);
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::findByExample Result -> {}", source);
+
 		// Convert entity to resource
 		return Optional.of(source.flatMap(
 				(input) -> (input == null) ? Optional.empty() : Optional.of(mapperToResource().toResource(input)))
@@ -155,7 +191,12 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 	 */
 	@Override
 	public Optional<R> findById(final Long id) throws NotFoundEntityException {
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::findById Finding a entry with id: {}", id);
+
 		final Optional<T> source = getEntityRepository().findById(id);
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::findById Result -> {}", source);
+
 		// Convert entity to resource
 		return Optional.of(source.flatMap(
 				(input) -> (input == null) ? Optional.empty() : Optional.of(mapperToResource().toResource(input)))
@@ -177,7 +218,9 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 		final Iterable<T> entities = StreamSupport.stream(sources.spliterator(), false)
 				.map(mapperToEntity()::apply)
 				.collect(Collectors.toList());
-		
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::saveAll Result -> {}", entities);
+
 		// Convert entity to resource
 		return StreamSupport.stream(entities.spliterator(), false)
 				.map(mapperToResource()::toResource)
@@ -191,10 +234,13 @@ public abstract class BaseEntityService<T extends BaseEntity, R extends BaseEnti
 	public R update(final R entity) throws NotFoundEntityException {
 		if (entity.getEntityId() == null)
 			throw new NotFoundEntityException();
-		
+
 		final T source = getEntityRepository().save(mapperToEntity().apply(entity));
+
+		LoggerUtil.message(LogMessageLevel.INFO, "BaseEntityService::update Result -> {}", source);
+
 		// Convert entity to resource
 		return mapperToResource().toResource(source);
 	}
-	
+
 }
